fix(moon-lander): fix remaining lint warnings
