+++ examples/moon-lander-react/package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "dev": "npx tsx scripts/dev.ts",
     "build": "vite build",
-    "test": "vitest run --config vitest.config.browser.ts"
+    "test": "vitest run --config vitest.config.browser.ts",
+    "walkthrough": "npx @marp-team/marp-cli --html walkthrough/jazz-moon-lander.md --preview",
+    "walkthrough:shots": "playwright test --config playwright.screenshots.config.ts"
   },
   "dependencies": {
     "jazz-tools": "workspace:*",
@@ -15,6 +17,7 @@
     "tsx": "^4.21.0"
   },
   "devDependencies": {
+    "@playwright/test": "^1.58.2",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
     "@vitejs/plugin-react": "^4.0.0",
